prototype: round 3 — budget grouping idioms + hover drag handles (#260)
Round-2 feedback: indentation alone doesn't tie accounts to their budget,
and the reorder drag handles had vanished. All slots keep their active
treatment and gain hover-reveal handles (visual-only) plus a grouping
idiom: A = vertical hairline tree guide, B = bordered surface slab per
budget, C = bend-arrow connectors. A second budget was seeded into the
worktree's dev-DB copy to make grouping judgeable.
